@@ -5,13 +5,13 @@ publish_to: none
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.10.1"
+  flutter: ">=3.24.1"
 
 dependencies:
   flutter_localizations:
     sdk: flutter
 
-  collection: ^1.15.0-nullsafety.4
+  collection: ^1.18.0
   deriv_technical_analysis:
     git:
       url: git@github.com:regentmarkets/flutter-deriv-packages.git
@@ -19,10 +19,10 @@ dependencies:
       ref: dev
 
   equatable: ^2.0.5
-  intl: ^0.18.0
-  json_annotation: ^4.8.0
-  provider: ^6.0.5
-  shared_preferences: ^2.1.0
+  intl: ^0.19.0
+  json_annotation: ^4.9.0
+  provider: ^6.1.2
+  shared_preferences: ^2.2.3
 
 dev_dependencies:
   flutter_test:
@@ -33,9 +33,9 @@ dev_dependencies:
       path: packages/deriv_lint
       ref: dev
 
-  build_runner: ^2.3.3
-  intl_utils: ^2.8.2
-  json_serializable: ^6.6.1
+  build_runner: ^2.4.9
+  intl_utils: ^2.8.7
+  json_serializable: ^6.8.0
 
 flutter:
   assets:
