@@ -104,7 +104,7 @@ class _FullscreenChartState extends State<FullscreenChart> {
   Asset _symbol = Asset(name: 'R_50');
 
   final ChartController _controller = ChartController();
-  PersistentBottomSheetController<dynamic>? _bottomSheetController;
+  PersistentBottomSheetController? _bottomSheetController;
 
   late PrefServiceCache _prefService;
 
@@ -219,7 +219,7 @@ class _FullscreenChartState extends State<FullscreenChart> {
 
   Future<void> _getActiveSymbols() async {
     _activeSymbols = (await ActiveSymbolsResponse.fetchActiveSymbols(
-      const ActiveSymbolsRequest(activeSymbols: 'brief', productType: 'basic'),
+      const ActiveSymbolsRequest(activeSymbols: 'brief', productType: 'basic', landingCompany: 'svg'),
     ))
         .activeSymbols!;
 
@@ -660,7 +660,7 @@ class _FullscreenChartState extends State<FullscreenChart> {
   Widget _buildMarketSelectorButton() => MarketSelectorButton(
         asset: _symbol,
         onTap: () {
-          _bottomSheetController = showBottomSheet<void>(
+          _bottomSheetController = showBottomSheet(
             backgroundColor: Colors.transparent,
             context: context,
             builder: (BuildContext context) => MarketSelector(
